telegram-integration/requirements.txt
click==7.1.2
cryptography==3.4.1
fastapi==0.63.0
getmac==0.8.2
h11==0.12.0
pycparser==2.20
pycryptodome==3.9.9
paho-mqtt==1.5.1
loguru==0.5.3
telethon==1.32.0
PyYAML==5.4.1

